Added simple WeatherApi scheduler

+++ src/main/java/pl/luki2183/farmManager/weatherInfo/model/WeatherInfoEntity.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 @Embeddable
 public class WeatherInfoEntity {
-    private Double humidity;
+    private Integer humidity;
     private Double windSpeed;
 }
